+++ src/wai/annotations/isp/label_present/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package containing the label-present ISP.
+"""