+++ src/wai/annotations/isp/label_present/specifier/__init__.py
@@ -1,0 +1,1 @@
+from ._LabelPresentISPSpecifier import LabelPresentISPSpecifier